feat(kmp): Step 5 — extract simulationModule + navigationModule into :core CoreModule.kt
Create core/src/commonMain/kotlin/.../di/CoreModule.kt with simulationCoreModule
(SimulationProcessFactory, GridTransformer) and navigationModule (all navigation
service scope bindings) — both using only commonMain classes, KMP-clean.

Update InterlockSimModule.kt to replace the inlined simulationModule + navigationModule
with includes(coreModule) + simulationDesktopModule (ContextTransformer,
DefaultSimulationContextFactory, ExampleRegistry — JVM-only bindings that remain
in desktop-ui). Update InterlockSimTestModule.kt accordingly.

Key fix: declare simulationCoreModule and navigationModule before coreModule in the
same file to avoid Kotlin top-level val JVM static initialization ordering issue
(null in includes() list).

All tests pass: :core jvmTest 1569, :desktop-ui:test 442.
