+++ desktop-ui/src/test/kotlin/cz/vutbr/fit/interlockSim/testutil/InterlockSimTestModule.kt
@@ -1,11 +1,11 @@
 package cz.vutbr.fit.interlockSim.testutil
 
+import cz.vutbr.fit.interlockSim.di.coreModule
 import cz.vutbr.fit.interlockSim.di.editingModule
 import cz.vutbr.fit.interlockSim.di.guiModule
 import cz.vutbr.fit.interlockSim.di.interlockSimModule
-import cz.vutbr.fit.interlockSim.di.navigationModule
 import cz.vutbr.fit.interlockSim.di.objectsModule
-import cz.vutbr.fit.interlockSim.di.simulationModule
+import cz.vutbr.fit.interlockSim.di.simulationDesktopModule
 import cz.vutbr.fit.interlockSim.di.utilModule
 import cz.vutbr.fit.interlockSim.di.xmlModule
 import org.koin.core.module.Module
@@ -26,8 +26,8 @@ val testModuleLightweight: Module =
 			objectsModule,
 			xmlModule,
 			editingModule,
-			simulationModule,
-			navigationModule
+			coreModule, // simulationCoreModule + navigationModule (commonMain, KMP-clean)
+			simulationDesktopModule // JVM-only: ContextTransformer, DefaultSimulationContextFactory, ExampleRegistry
 			// NOTE: guiModule excluded to prevent Frame overhead
 		)
 		// Provide a new instance of TestContextBuilder for each injection
@@ -63,8 +63,8 @@ val integrationTestModule: Module =
 			objectsModule,
 			xmlModule,
 			editingModule,
-			simulationModule,
-			navigationModule
+			coreModule, // simulationCoreModule + navigationModule (commonMain, KMP-clean)
+			simulationDesktopModule // JVM-only: ContextTransformer, DefaultSimulationContextFactory, ExampleRegistry
 			// NOTE: guiModule excluded to prevent Frame overhead
 		)
 		// Provide a new instance of TestContextBuilder for each injection
